+++ src/django/api/models/utility.py
@@ -1,8 +1,10 @@
 from django.db import models
 from django.contrib.gis.db import models as gis_models
 from django.contrib.gis.geos import Point
+from .state import State
 
 RALEIGH = Point(-78.6382, 35.7796)
+
 
 __all__ = ["Utility"]
 
@@ -11,6 +13,7 @@ class Utility(models.Model):
     pwsid = models.CharField(max_length=9, unique=True)
     name = models.CharField(max_length=127)
     location = gis_models.PointField(geography=True, default=RALEIGH)
+    state = models.ForeignKey(State, on_delete=models.PROTECT, default="NC")
 
     class Meta:
         verbose_name_plural = "utilities"